FIX-3006 Add EtpDataWorkOrderService
